ai-rag/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
chromadb>=0.5.0
requests>=2.31.0
pyyaml>=6.0.1
pydantic>=2.7.0
python-dotenv>=1.0.1
